pyproject.toml
[tool.poetry]
name = "lms-system"
version = "0.1.0"
description = ""
authors = ["Tamir Mandreev <tamirmandreev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
python-dotenv = "^1.1.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
coverage = "^7.8.0"
drf-yasg = "^1.21.10"
stripe = "^12.0.0"
celery = "^5.5.1"
redis = "^5.2.1"
django-celery-beat = "^2.8.0"
black = "^25.1.0"
isort = "^6.0.1"
gunicorn = "^23.0.0"
django-redis = "^5.4.0"
pytest = "^8.3.5"
pytest-django = "^4.11.1"
flake8 = "^7.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'config.settings'
python_files = ['tests.py']
addopts = '-p no:django_tests'
